Moves ButtonPrimary to its own directory
