+++ lib/ButtonNew/ButtonPrimary/ButtonPrimary.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import cx from 'classnames';
-import { ButtonBase } from './ButtonBase';
-import { propTypes, defaultProps, sizes } from './common';
+import { ButtonBase } from '../ButtonBase';
+import { propTypes, defaultProps, sizes } from '../common';
 
 function ButtonPrimary({
   children,
