+++ lib/ButtonNew/stories/ButtonPrimary.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { map } from 'lodash';
-import { ButtonPrimary } from '../ButtonPrimary';
+import { ButtonPrimary } from '../ButtonPrimary/ButtonPrimary';
 import StoryItem from '../../../stories/styleguide/StoryItem';
 import { ButtonStoryItem } from './ButtonStoryItem';
