+++ lib/index.js
@@ -121,4 +121,4 @@ export { Windowing } from './Windowing';
 export FinderPanelLayout from './FinderPanelLayout';
 export { PillInput } from './PillInput/PillInput';
 export DismissiblePrompt from './DismissiblePrompt';
-export { ButtonPrimary } from './ButtonNew/ButtonPrimary';
+export { ButtonPrimary } from './ButtonNew/ButtonPrimary/ButtonPrimary';
